deletes must be confirmed

+++ src/presenters/cardModalPresenter.js
@@ -49,7 +49,11 @@ export default function CardModalPresenter(props) {
           props.model.moveCard(props.card, props.boardId, newBoardId);
         }}
         onDeleteCard={() => {
-          props.model.deleteCard(props.card.id, props.boardId);
+          if (window.confirm("Are you sure you want to delete this card?")) {
+            props.model.deleteCard(props.card.id, props.boardId);
+          } else {
+            return;
+          }
         }}
         setTag={(newTag) => {
           setTagText(newTag);
